requirements.txt
django==1.11.7
psycopg2==2.7.3.2
Pillow==4.3.0
django-debug-toolbar==1.9.1
python-bittrex==0.2.1
django-suit==0.2.25
django-suit-locale==1.0.10
# django-admin-bootstrapped==2.5.7
numpy
TA-Lib
raven==6.4.0
telebot
# python-binance==0.7.0
python-binance==0.7.5
